requirements.txt
torch>=1.10.0
transformers>=4.20.0
tokenizers>=0.10.0
scikit-learn>=0.24.0
numpy>=1.20.0
matplotlib>=3.5.0
wandb
python-dotenv
datasets
accelerate>=0.26.0
peft